Add files via upload
The results and discussion section is a critical part of any research paper as it presents the findings of the study and provides an in-depth analysis and interpretation of the results. In this section, we will discuss the outcomes of the proposed system, including the performance of the XGBoost model trained on retinal biometric data for cybersecurity authentication.